# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/decks/import": {
      "post": {
        "operationId": "CardsController_importCards",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCardsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/cards/move": {
      "post": {
        "operationId": "CardsController_move",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveCardsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/decks/{deckId}/cards": {
      "get": {
        "operationId": "CardsController_list",
        "parameters": [
          {
            "name": "deckId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      },
      "post": {
        "operationId": "CardsController_create",
        "parameters": [
          {
            "name": "deckId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/decks/{deckId}/cards/{cardId}": {
      "patch": {
        "operationId": "CardsController_update",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deckId",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      },
      "delete": {
        "operationId": "CardsController_remove",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deckId",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/cards/{cardId}/images": {
      "post": {
        "operationId": "CardsController_attachImage",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/cards/{cardId}/images/{imageId}": {
      "delete": {
        "operationId": "CardsController_removeImage",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cards"
        ]
      }
    },
    "/media/{path}": {
      "get": {
        "operationId": "MediaController_getImage",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
    "/devices/register": {
      "post": {
        "operationId": "registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Devices"
        ]
      }
    },
    "/devices/{deviceId}": {
      "patch": {
        "operationId": "updateDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_delete",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      }
    },
...
  },
  "info": {
    "title": "iTu API",
    "description": "Authoritative REST contract for iTu productivity & learning system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ImportCardsDto": {
        "type": "object",
        "properties": {}
      },
      "MoveCardsDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCardDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCardDto": {
        "type": "object",
        "properties": {}
      },
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDeviceDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}